Moved RenderingServiceException to Gateway
