+++ components/blitz/src/omero/gateway/exception/RenderingServiceException.java
@@ -21,7 +21,7 @@
  *------------------------------------------------------------------------------
  */
 
-package org.openmicroscopy.shoola.env.rnd;
+package omero.gateway.exception;
 
 
 //Java imports
